+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased](https://github.com/dalance/procs/compare/v0.11.6...Unreleased) - ReleaseDate
 
+* [Added] elapsed time [#120](https://github.com/dalance/procs/issues/120)
 * [Added] completion file message [#130](https://github.com/dalance/procs/issues/130)
 
 ## [v0.11.6](https://github.com/dalance/procs/compare/v0.11.5...v0.11.6) - 2021-05-18